+++ packages/stream_chat_flutter/lib/src/theme/stream_chat_theme.dart
@@ -57,7 +57,6 @@ class StreamChatThemeData {
     StreamPollCommentsSheetThemeData? pollCommentsSheetTheme,
     StreamPollOptionVotesSheetThemeData? pollOptionVotesSheetTheme,
     StreamThreadListTileThemeData? threadListTileTheme,
-    StreamDraftListTileThemeData? draftListTileTheme,
     StreamVoiceRecordingAttachmentThemeData? voiceRecordingAttachmentTheme,
     StreamChannelListItemThemeData? channelListItemTheme,
   }) {
@@ -86,7 +85,6 @@ class StreamChatThemeData {
       pollCommentsSheetTheme: pollCommentsSheetTheme,
       pollOptionVotesSheetTheme: pollOptionVotesSheetTheme,
       threadListTileTheme: threadListTileTheme,
-      draftListTileTheme: draftListTileTheme,
       voiceRecordingAttachmentTheme: voiceRecordingAttachmentTheme,
       channelListItemTheme: channelListItemTheme,
     );
@@ -117,7 +115,6 @@ class StreamChatThemeData {
     required this.pollCommentsSheetTheme,
     required this.pollOptionVotesSheetTheme,
     required this.threadListTileTheme,
-    required this.draftListTileTheme,
     required this.voiceRecordingAttachmentTheme,
     required this.channelListItemTheme,
   });
@@ -194,19 +191,6 @@ class StreamChatThemeData {
       pollCommentsSheetTheme: const StreamPollCommentsSheetThemeData(),
       pollOptionVotesSheetTheme: const StreamPollOptionVotesSheetThemeData(),
       threadListTileTheme: const StreamThreadListTileThemeData(),
-      draftListTileTheme: StreamDraftListTileThemeData(
-        backgroundColor: colorTheme.barsBg,
-        padding: const EdgeInsets.symmetric(vertical: 14, horizontal: 8),
-        draftChannelNameStyle: textTheme.bodyBold.copyWith(
-          color: colorTheme.textHighEmphasis,
-        ),
-        draftMessageStyle: textTheme.footnote.copyWith(
-          color: colorTheme.textLowEmphasis,
-        ),
-        draftTimestampStyle: textTheme.footnote.copyWith(
-          color: colorTheme.textLowEmphasis,
-        ),
-      ),
       voiceRecordingAttachmentTheme: const StreamVoiceRecordingAttachmentThemeData(),
       channelListItemTheme: const StreamChannelListItemThemeData(),
     );
@@ -265,9 +249,6 @@ class StreamChatThemeData {
   /// Theme configuration for the [StreamChannelListItem] widget.
   final StreamChannelListItemThemeData channelListItemTheme;
 
-  /// Theme configuration for the [StreamDraftListTile] widget.
-  final StreamDraftListTileThemeData draftListTileTheme;
-
   /// Creates a copy of [StreamChatThemeData] with specified attributes
   /// overridden.
   StreamChatThemeData copyWith({
@@ -288,7 +269,6 @@ class StreamChatThemeData {
     StreamPollCommentsSheetThemeData? pollCommentsSheetTheme,
     StreamPollOptionVotesSheetThemeData? pollOptionVotesSheetTheme,
     StreamThreadListTileThemeData? threadListTileTheme,
-    StreamDraftListTileThemeData? draftListTileTheme,
     StreamVoiceRecordingAttachmentThemeData? voiceRecordingAttachmentTheme,
     StreamChannelListItemThemeData? channelListItemTheme,
   }) => StreamChatThemeData.raw(
@@ -307,7 +287,6 @@ class StreamChatThemeData {
     pollCommentsSheetTheme: pollCommentsSheetTheme ?? this.pollCommentsSheetTheme,
     pollOptionVotesSheetTheme: pollOptionVotesSheetTheme ?? this.pollOptionVotesSheetTheme,
     threadListTileTheme: threadListTileTheme ?? this.threadListTileTheme,
-    draftListTileTheme: draftListTileTheme ?? this.draftListTileTheme,
     voiceRecordingAttachmentTheme: voiceRecordingAttachmentTheme ?? this.voiceRecordingAttachmentTheme,
     channelListItemTheme: channelListItemTheme ?? this.channelListItemTheme,
   );
@@ -331,7 +310,6 @@ class StreamChatThemeData {
       pollCommentsSheetTheme: pollCommentsSheetTheme.merge(other.pollCommentsSheetTheme),
       pollOptionVotesSheetTheme: pollOptionVotesSheetTheme.merge(other.pollOptionVotesSheetTheme),
       threadListTileTheme: threadListTileTheme.merge(other.threadListTileTheme),
-      draftListTileTheme: draftListTileTheme.merge(other.draftListTileTheme),
       voiceRecordingAttachmentTheme: voiceRecordingAttachmentTheme.merge(other.voiceRecordingAttachmentTheme),
       channelListItemTheme: channelListItemTheme.merge(other.channelListItemTheme),
     );