+++ packages/stream_chat_flutter/lib/src/theme/themes.dart
@@ -2,7 +2,6 @@ export 'avatar_theme.dart';
 export 'channel_header_theme.dart';
 export 'channel_list_header_theme.dart';
 export 'color_theme.dart';
-export 'draft_list_tile_theme.dart';
 export 'gallery_footer_theme.dart';
 export 'gallery_header_theme.dart';
 export 'message_list_view_theme.dart';