+++ sample_app/lib/widgets/stream_draft_list_view.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:stream_chat_flutter/src/scroll_view/stream_scroll_view_error_widget.dart';
-import 'package:stream_chat_flutter/src/scroll_view/stream_scroll_view_load_more_error.dart';
-import 'package:stream_chat_flutter/src/scroll_view/stream_scroll_view_loading_widget.dart';
+import 'package:sample_app/widgets/stream_draft_list_tile.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:stream_core_flutter/stream_core_flutter.dart';
 
@@ -17,7 +15,6 @@ Widget defaultDraftListViewSeparatorBuilder(
 /// [StreamDraftListView].
 typedef StreamDraftListViewIndexedWidgetBuilder = StreamScrollViewIndexedWidgetBuilder<Draft, StreamDraftListTile>;
 
-/// {@template streamDraftListView}
 /// A [ListView] that shows a list of [Draft]'s. It uses a
 /// [StreamDraftListController] to load the drafts in paginated form.
 ///
@@ -38,9 +35,8 @@ typedef StreamDraftListViewIndexedWidgetBuilder = StreamScrollViewIndexedWidgetB
 /// See also:
 /// * [StreamDraftListTile]
 /// * [StreamDraftListController]
-/// {@endtemplate}
 class StreamDraftListView extends StatelessWidget {
-  /// {@macro streamDraftListView}
+  /// Creates a new [StreamDraftListView].
   const StreamDraftListView({
     super.key,
     required this.controller,
@@ -96,11 +92,9 @@ class StreamDraftListView extends StatelessWidget {
   /// The index to take into account when triggering [controller.loadMore].
   final int loadMoreTriggerIndex;
 
-  /// {@template flutter.widgets.scroll_view.scrollDirection}
   /// The axis along which the scroll view scrolls.
   ///
   /// Defaults to [Axis.vertical].
-  /// {@endtemplate}
   final Axis scrollDirection;
 
   /// The amount of space by which to inset the children.
@@ -108,149 +102,41 @@ class StreamDraftListView extends StatelessWidget {
 
   /// Whether to wrap each child in an [AutomaticKeepAlive].
   ///
-  /// Typically, children in lazy list are wrapped in [AutomaticKeepAlive]
-  /// widgets so that children can use [KeepAliveNotification]s to preserve
-  /// their state when they would otherwise be garbage collected off-screen.
-  ///
-  /// This feature (and [addRepaintBoundaries]) must be disabled if the children
-  /// are going to manually maintain their [KeepAlive] state. It may also be
-  /// more efficient to disable this feature if it is known ahead of time that
-  /// none of the children will ever try to keep themselves alive.
-  ///
   /// Defaults to true.
   final bool addAutomaticKeepAlives;
 
   /// Whether to wrap each child in a [RepaintBoundary].
-  ///
-  /// Typically, children in a scrolling container are wrapped in repaint
-  /// boundaries so that they do not need to be repainted as the list scrolls.
-  /// If the children are easy to repaint (e.g., solid color blocks or a short
-  /// snippet of text), it might be more efficient to not add a repaint boundary
-  /// and simply repaint the children during scrolling.
   ///
   /// Defaults to true.
   final bool addRepaintBoundaries;
 
   /// Whether to wrap each child in an [IndexedSemantics].
   ///
-  /// Typically, children in a scrolling container must be annotated with a
-  /// semantic index in order to generate the correct accessibility
-  /// announcements. This should only be set to false if the indexes have
-  /// already been provided by an [IndexedSemantics] widget.
-  ///
   /// Defaults to true.
-  ///
-  /// See also:
-  ///
-  ///  * [IndexedSemantics], for an explanation of how to manually
-  ///    provide semantic indexes.
   final bool addSemanticIndexes;
 
-  /// {@template flutter.widgets.scroll_view.reverse}
   /// Whether the scroll view scrolls in the reading direction.
   ///
-  /// For example, if [scrollDirection] is [Axis.vertical], then the scroll view
-  /// scrolls from top to bottom when [reverse] is false and from bottom to top
-  /// when [reverse] is true.
-  ///
   /// Defaults to false.
-  /// {@endtemplate}
   final bool reverse;
 
-  /// {@template flutter.widgets.scroll_view.controller}
   /// An object that can be used to control the position to which this scroll
   /// view is scrolled.
-  ///
-  /// Must be null if [primary] is true.
-  ///
-  /// A [ScrollController] serves several purposes. It can be used to control
-  /// the initial scroll position (see [ScrollController.initialScrollOffset]).
-  /// It can be used to control whether the scroll view should automatically
-  /// save and restore its scroll position in the [PageStorage] (see
-  /// [ScrollController.keepScrollOffset]). It can be used to read the current
-  /// scroll position (see [ScrollController.offset]), or change it (see
-  /// [ScrollController.animateTo]).
-  /// {@endtemplate}
   final ScrollController? scrollController;
 
-  /// {@template flutter.widgets.scroll_view.primary}
   /// Whether this is the primary scroll view associated with the parent
   /// [PrimaryScrollController].
-  ///
-  /// When this is true, the scroll view is scrollable even if it does not have
-  /// sufficient content to actually scroll. Otherwise, by default the user can
-  /// only scroll the view if it has sufficient content. See [physics].
-  ///
-  /// Also when true, the scroll view is used for default [ScrollAction]s. If a
-  /// ScrollAction is not handled by an otherwise focused part of the
-  /// application, the ScrollAction will be evaluated using this scroll view,
-  /// for example, when executing [Shortcuts] key events like page up and down.
-  ///
-  /// On iOS, this also identifies the scroll view that will scroll to top in
-  /// response to a tap in the status bar.
-  /// {@endtemplate}
   ///
   /// Defaults to true when [scrollController] is null.
   final bool? primary;
 
-  /// {@template flutter.widgets.scroll_view.shrinkWrap}
   /// Whether the extent of the scroll view in the [scrollDirection] should be
   /// determined by the contents being viewed.
   ///
-  /// If the scroll view does not shrink wrap, then the scroll view will expand
-  /// to the maximum allowed size in the [scrollDirection]. If the scroll view
-  /// has unbounded constraints in the [scrollDirection], then [shrinkWrap] must
-  /// be true.
-  ///
-  /// Shrink wrapping the content of the scroll view is significantly more
-  /// expensive than expanding to the maximum allowed size because the content
-  /// can expand and contract during scrolling, which means the size of the
-  /// scroll view needs to be recomputed whenever the scroll position changes.
-  ///
   /// Defaults to false.
-  /// {@endtemplate}
   final bool shrinkWrap;
 
-  /// {@template flutter.widgets.scroll_view.physics}
   /// How the scroll view should respond to user input.
-  ///
-  /// For example, determines how the scroll view continues to animate after the
-  /// user stops dragging the scroll view.
-  ///
-  /// Defaults to matching platform conventions. Furthermore, if [primary] is
-  /// false, then the user cannot scroll if there is insufficient content to
-  /// scroll, while if [primary] is true, they can always attempt to scroll.
-  ///
-  /// To force the scroll view to always be scrollable even if there is
-  /// insufficient content, as if [primary] was true but without necessarily
-  /// setting it to true, provide an [AlwaysScrollableScrollPhysics] physics
-  /// object, as in:
-  ///
-  /// ```dart
-  ///   physics: const AlwaysScrollableScrollPhysics(),
-  /// ```
-  ///
-  /// To force the scroll view to use the default platform conventions and not
-  /// be scrollable if there is insufficient content, regardless of the value of
-  /// [primary], provide an explicit [ScrollPhysics] object, as in:
-  ///
-  /// ```dart
-  ///   physics: const ScrollPhysics(),
-  /// ```
-  ///
-  /// The physics can be changed dynamically (by providing a new object in a
-  /// subsequent build), but new physics will only take effect if the _class_ of
-  /// the provided object changes. Merely constructing a new instance with a
-  /// different configuration is insufficient to cause the physics to be
-  /// reapplied. (This is because the final object used is generated
-  /// dynamically, which can be relatively expensive, and it would be
-  /// inefficient to speculatively create this object each frame to see if the
-  /// physics should be updated.)
-  /// {@endtemplate}
-  ///
-  /// If an explicit [ScrollBehavior] is provided to [scrollBehavior], the
-  /// [ScrollPhysics] provided by that behavior will take precedence after
-  /// [physics].
   final ScrollPhysics? physics;
 
   /// {@macro flutter.rendering.RenderViewportBase.cacheExtent}
@@ -259,10 +145,7 @@ class StreamDraftListView extends StatelessWidget {
   /// {@macro flutter.widgets.scrollable.dragStartBehavior}
   final DragStartBehavior dragStartBehavior;
 
-  /// {@template flutter.widgets.scroll_view.keyboardDismissBehavior}
-  /// [ScrollViewKeyboardDismissBehavior] the defines how this [ScrollView] will
-  /// dismiss the keyboard automatically.
-  /// {@endtemplate}
+  /// Defines how this [ScrollView] will dismiss the keyboard automatically.
   final ScrollViewKeyboardDismissBehavior keyboardDismissBehavior;
 
   /// {@macro flutter.widgets.scrollable.restorationId}
@@ -317,29 +200,89 @@ class StreamDraftListView extends StatelessWidget {
               emptyTitle: Text(context.translations.emptyMessagesText),
             ),
           ),
-      loadMoreErrorBuilder: (context, error) => StreamScrollViewLoadMoreError.list(
+      loadMoreErrorBuilder: (context, error) => _LoadMoreError(
         onTap: controller.retry,
-        error: Text(context.translations.loadingMessagesError),
+        error: context.translations.loadingMessagesError,
       ),
-      loadMoreIndicatorBuilder: (context) => Center(
+      loadMoreIndicatorBuilder: (context) => const Center(
         child: Padding(
-          padding: const EdgeInsets.all(16),
-          child: StreamLoadingSpinner(),
+          padding: EdgeInsets.all(16),
+          child: SizedBox.square(
+            dimension: 24,
+            child: CircularProgressIndicator.adaptive(),
+          ),
         ),
       ),
       loadingBuilder: (context) =>
           loadingBuilder?.call(context) ??
           const Center(
-            child: StreamScrollViewLoadingWidget(),
+            child: SizedBox.square(
+              dimension: 42,
+              child: CircularProgressIndicator.adaptive(),
+            ),
           ),
       errorBuilder: (context, error) =>
           errorBuilder?.call(context, error) ??
           Center(
-            child: StreamScrollViewErrorWidget(
-              errorTitle: Text(context.translations.loadingMessagesError),
-              onRetryPressed: controller.refresh,
+            child: _ErrorWidget(
+              errorTitle: context.translations.loadingMessagesError,
+              onRetry: controller.refresh,
             ),
           ),
+    );
+  }
+}
+
+/// A simple error widget shown when the draft list fails to load.
+class _ErrorWidget extends StatelessWidget {
+  const _ErrorWidget({
+    required this.errorTitle,
+    required this.onRetry,
+  });
+
+  final String errorTitle;
+  final VoidCallback onRetry;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Text(errorTitle),
+        const SizedBox(height: 8),
+        TextButton(
+          onPressed: onRetry,
+          child: const Text('Retry'),
+        ),
+      ],
+    );
+  }
+}
+
+/// A simple load-more error widget shown inline in the list.
+class _LoadMoreError extends StatelessWidget {
+  const _LoadMoreError({
+    required this.onTap,
+    required this.error,
+  });
+
+  final VoidCallback onTap;
+  final String error;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onTap,
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(error, style: const TextStyle(color: Colors.white)),
+            const Icon(Icons.refresh, color: Colors.white),
+          ],
+        ),
+      ),
     );
   }
 }